# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Система оценок (Gradebook API)",
        "description": "API для управления студентами, предметами и их оценками. Использованы Laravel Resources и hasManyThrough.",
        "contact": {
            "email": "artemtaiti@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel API Server"
        }
    ],
    "paths": {
...
        "/students": {
            "get": {
                "tags": [
                    "Студенты"
                ],
                "summary": "Получить список всех студентов",
                "operationId": "602b44d362d13bdd9b9131f32a937ae6",
                "responses": {
                    "200": {
                        "description": "Успешный ответ. Возвращает коллекцию студентов.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/StudentResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Студенты"
                ],
                "summary": "Создать нового студента",
                "operationId": "07c462c73c16ecd99c5f5b2e67339e23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "student_id_number"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Сергей"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Петров"
                                    },
                                    "student_id_number": {
                                        "type": "string",
                                        "example": "S1002"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Студент успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/StudentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/students/{student}": {
            "get": {
                "tags": [
                    "Студенты"
                ],
                "summary": "Получить данные студента по ID",
                "description": "Включает в себя Eager Loading связей grades и subjects.",
                "operationId": "5b0d7aa428bdb9904461f87fb75c9570",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "ID студента",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/StudentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Студент не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Студенты"
                ],
                "summary": "Обновить данные студента",
                "operationId": "dc5a12f258e57ade4154d0b6e99b212d",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "ID студента",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "student_id_number"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Сергей (Обновлено)"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Петров"
                                    },
                                    "student_id_number": {
                                        "type": "string",
                                        "example": "S1002"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное обновление",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/StudentResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Студент не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Студенты"
                ],
                "summary": "Удалить студента",
                "operationId": "24acd7e6964cb72096ada31630c5051e",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "description": "ID студента",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Студент успешно удален (Нет содержимого)"
                    },
                    "404": {
                        "description": "Студент не найден"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GradeResource": {
                "title": "Grade Resource (Output)",
                "description": "Оценка, возвращаемая API, включая связанный предмет",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "student_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "score": {
                        "type": "integer",
                        "example": 5
                    },
                    "exam_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-11-20"
                    },
                    "subject": {
                        "description": "Связанный предмет",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 3
                            },
                            "title": {
                                "type": "string",
                                "example": "Высшая математика"
                            },
                            "lecturer_name": {
                                "type": "string",
                                "example": "Петров"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StudentResource": {
                "title": "Student Resource (Output)",
                "description": "Студент, возвращаемый API, включая оценки и предметы.",
                "properties": {
                    "id": {
                        "description": "ID студента",
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "description": "Имя",
                        "type": "string",
                        "example": "Иван"
                    },
                    "last_name": {
                        "description": "Фамилия",
                        "type": "string",
                        "example": "Иванов"
                    },
                    "student_id_number": {
                        "description": "Номер зачетной книжки / ID",
                        "type": "string",
                        "example": "S1001"
                    },
                    "grades": {
                        "description": "Список оценок студента",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GradeResource"
                        }
                    },
                    "subjects": {
                        "description": "Список уникальных предметов, по которым есть оценки (из hasManyThrough)",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 3
                                },
                                "title": {
                                    "type": "string",
                                    "example": "Высшая математика"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}